save new withdrawals

+++ db/migrations/up/1506434516_users.up.sql
@@ -4,10 +4,10 @@ CREATE TABLE users
     CONSTRAINT users_pkey
     PRIMARY KEY,
   trade_link_url  VARCHAR(120),
-  bitcoin_balance INTEGER DEFAULT 0  NOT NULL,
-  csgokey_balance INTEGER DEFAULT 0  NOT NULL,
-  display_name    VARCHAR(40)        NOT NULL,
-  avatar          VARCHAR(255)       NOT NULL,
+  bitcoin_balance NUMERIC(10, 8) DEFAULT 0  NOT NULL,
+  csgokey_balance INTEGER DEFAULT 0         NOT NULL,
+  display_name    VARCHAR(40)               NOT NULL,
+  avatar          VARCHAR(255)              NOT NULL,
   added_at        TIMESTAMP WITH TIME ZONE DEFAULT now()
 );
